sửa lại db với dữ liệu mới và show toàn bộ DoanVien

+++ BE/src/models/phieuthudoanphi.model.js
@@ -22,7 +22,7 @@ const PhieuThuDoanPhi = sequelize.define(
       allowNull: true,
     },
     tongTien: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(18, 2),
       allowNull: true,
     },
     fileDinhKem: {
